[refactor] UploadImg 컴포넌트 위치 변경 (common -> making)
