+++ src/components/making/index.js
@@ -8,5 +8,6 @@ export { default as SubTitle, BtnIcon } from "./SubTitle";
 export { default as Options } from "./Option";
 
 // component
+export { default as UploadImg } from "./UploadImg";
 export { default as Questions } from "./Question";
 export { default as Result } from "./Result";
